Not running! Sorry! No warnings on nix, but +1 segfaults

+++ src/gecom/UnitSquare.hpp
@@ -11,16 +11,16 @@ namespace gecom {
 		GLuint vaoID;
 		GLuint vboID;
 	public:
-		UnitSquare(std::shared_ptr<Entity> parent, double half_width, double half_height) : DrawableComponent(parent) {
+		UnitSquare(std::shared_ptr<Entity> parent, float half_width, float half_height) : DrawableComponent(parent) {
 			glGenVertexArrays(1, &vaoID);
 			glBindVertexArray(vaoID);
 			glGenBuffers(1, &vboID);
 
 			float pos[] = {
-				-1 * half_width, -1 * half_height, 0,
-				-1 * half_width, 1 * half_height, 0,
-				1 * half_width, -1 * half_height, 0,
-				1 * half_width, 1 * half_height, 0
+				-1.0f * half_width, -1.0f * half_height, 0.0f,
+				-1.0f * half_width, 1.0f * half_height, 0.0f,
+				1.0f * half_width, -1.0f * half_height, 0.0f,
+				1.0f * half_width, 1.0f * half_height, 0.0f
 			};
 			glBindBuffer(GL_ARRAY_BUFFER, vboID);
 			glBufferData(GL_ARRAY_BUFFER, 3 * 4 * sizeof(float), pos, GL_STATIC_DRAW);
